Move keeping selected user to store
